+++ lib/index.js
@@ -1,12 +1,12 @@
-// dependencies
 var _ = require('lodash');
 var stylelint = require('stylelint');
 var doiuse = require('doiuse');
 var Result = require('postcss/lib/result');
-var cleanWarningText = require('./cleanWarningText');
-var parseOptions = require('./parseOptions');
 
-// plugin settings
+/**
+ * Plugin settings
+ */
+
 var ruleName = 'plugin/no-unsupported-browser-features';
 var messages = stylelint.utils.ruleMessages(ruleName, {
   rejected: function rejectRule(details) {
@@ -14,13 +14,58 @@ var messages = stylelint.utils.ruleMessages(ruleName, {
   }
 });
 
-// the allowed options
+/**
+ * The allowed options
+ */
+
 var optionsSchema = {
   browsers: [_.isString],
   ignore: [_.isString]
 };
 
-// the main plugin rule
+/**
+ * Utilities
+ */
+
+function cleanWarningText(warningText) {
+  // Get index of feature Id
+  var featureIdIndex = warningText.lastIndexOf('(');
+
+  // Get feature Id, then replace brackets with quotes
+  var featureId = warningText.slice(featureIdIndex, warningText.length).replace(/\(|\)/g, '"');
+
+  // Get start of support text i.e. "x not supported by...", or "y only partially supported by..."
+  var browserSupportStartIndex =
+    warningText.indexOf('not') !== -1 ? warningText.indexOf('not') : warningText.indexOf('only');
+
+  // Get browser support text, then strip brackets.
+  var browserSupport = warningText
+    .slice(browserSupportStartIndex, featureIdIndex - 1)
+    .replace(/\(|\)|:/g, '');
+
+  var cleanedWarningText = featureId + ' is ' + browserSupport;
+
+  return cleanedWarningText;
+}
+
+function parseOptions(options) {
+  var doiuseOptions = {};
+
+  if (options && options.browsers) {
+    doiuseOptions.browsers = options.browsers;
+  }
+
+  if (options && options.ignore) {
+    doiuseOptions.ignore = options.ignore;
+  }
+
+  return doiuseOptions;
+}
+
+/**
+ * The main plugin rule
+ */
+
 function ruleFunction(on, options) {
   return function postCssPlugin(root, result) {
     var validOptions = stylelint.utils.validateOptions(result, ruleName, {
@@ -48,9 +93,6 @@ function ruleFunction(on, options) {
   };
 }
 
-// main plugin export
 module.exports = stylelint.createPlugin(ruleName, ruleFunction);
-
-// other exports
 module.exports.ruleName = ruleName;
 module.exports.messages = messages;
